pyproject.toml
[project]
name = "lol-pick-ml"
version = "0.1.0"
description = ""
authors = [
    {name = "Gaibler",email = "patrick.giessler@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "tensorflow>=2.16.0,<2.20.0",
    "keras>=3.10.0,<4.0.0",
    "requests>=2.32.4,<3.0.0",
    "minio>=7.2.15,<8.0.0",
    "urllib3>=2.5.0,<3.0.0",
    "pydantic>=2.11.7,<3.0.0",
    "pika>=1.3.2,<2.0.0",
    "fastapi>=0.115.13,<0.116.0",
    "numpy>=1.26.0,<2.2.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "uvicorn>=0.30.0,<1.0.0",
    "psutil>=5.9.0,<6.0.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry]
package-mode = false